Rework
Keep Download button in frame with chart, top left (where it is now)
Add CEMD logo to header of CSV file (top right)
Add another footnote indicating that the list is sorted alpha, by publisher
Explore pulling the footnotes into the same container with the chart so that they are not lost in scrolls.
Add footnotes to CSV file: ideally, these would be on a tab 2 that has qualifying information, OR as footnotes under the chart.
Check to make sure there is no shadow behind the CEMD logo in the frame.